# Optimize Source#read_until method