@naitoh 2024-06-02
> BTW, can we remove `.elements.each("root/child") {|_|}` from
> 
> https://github.com/ruby/rexml/blob/3e3893d48357c04c4f3a7088819880905a64742d/benchmark/parse.yaml#L50
> ?
> 
> I think that "dom" is slower than others because of it. And `.elements.each("root/child") {|_|}` isn't parse processing.

@kou 

I agree.
I created https://github.com/ruby/rexml/pull/136.
Thanks!